[Fix] preference 기능 수정
