+++ app/src/main/java/com/terbuck/terbuck/api/TokenManager.kt
@@ -48,31 +48,6 @@ class TokenManager(val context: Context) {
         return sharedPreferences.getString("school", "")
     }
 
-    fun saveIsSignUp(isSignUp: Boolean) {
-        val editor = sharedPreferences.edit()
-        editor.putBoolean("isSignUp", isSignUp)
-        editor.apply()
-    }
-
-    fun getIsSignUp(): Boolean {
-        return sharedPreferences.getBoolean("isSignUp", true)
-    }
-
-    fun saveIsUniversityRegistered(isRegistered: Boolean, isRequestRegistered: Boolean) {
-        val editor = sharedPreferences.edit()
-        editor.putBoolean("isRegistered", isRegistered)
-        editor.putBoolean("isRequestRegistered", isRequestRegistered)
-        editor.apply()
-    }
-
-    fun getIsUniversityRegistered(): Boolean {
-        return sharedPreferences.getBoolean("isRegistered", false)
-    }
-
-    fun getIsUniversityRequestRegistered(): Boolean {
-        return sharedPreferences.getBoolean("isRequestRegistered", false)
-    }
-
     fun clearAll() {
         val editor = sharedPreferences.edit()
         editor.clear()
